requirements.txt
flask==3.0.0
flask-sqlalchemy==3.0.5
pyjwt==2.8.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.1
psycopg2-binary==2.9.9
flask-cors==4.0.0
bcrypt==4.1.2
sqlalchemy==2.0.23
